♻️ Extract dropdown as component (#1706)
