+++ src/lib/components/TaskTables/TaskTableBodyCell.svelte
@@ -1,17 +1,11 @@
 <script lang="ts">
-  import { getStores } from '$app/stores';
-  import { enhance } from '$app/forms';
-
-  import { Dropdown, DropdownUl, DropdownLi, uiHelpers } from 'svelte-5-ui-lib';
   import EllipsisVertical from 'lucide-svelte/icons/ellipsis-vertical';
 
   import type { TaskResult } from '$lib/types/task';
 
   import GradeLabel from '$lib/components/GradeLabel.svelte';
   import ExternalLinkWrapper from '$lib/components/ExternalLinkWrapper.svelte';
-  import InputFieldWrapper from '$lib/components/InputFieldWrapper.svelte';
-
-  import { submission_statuses } from '$lib/services/submission_status';
+  import UpdatingDropdown from '$lib/components/SubmissionStatus/UpdatingDropdown.svelte';
 
   import { getTaskUrl, removeTaskIndexFromTitle } from '$lib/utils/task';
 
@@ -23,55 +17,7 @@
 
   let { taskResult, isLoggedIn, onupdate = () => {} }: Props = $props();
 
-  const { page } = getStores();
-  let activeUrl = $state($page.url.pathname);
-
-  let dropdown = uiHelpers();
-  let dropdownStatus = $state(false);
-  let closeDropdown = dropdown.close;
-
-  $effect(() => {
-    activeUrl = $page.url.pathname;
-    dropdownStatus = dropdown.isOpen;
-  });
-
-  let selectedSubmissionStatus = $state<{
-    innerId: string;
-    innerName: string;
-    labelName: string;
-  }>();
-  let showForm = $state(false);
-
-  function handleSubmit(submissionStatus: {
-    innerId: string;
-    innerName: string;
-    labelName: string;
-  }): void {
-    selectedSubmissionStatus = submissionStatus;
-    showForm = true;
-
-    // Submit after the form is rendered.
-    setTimeout(() => {
-      const submitButton = document.querySelector(
-        '#submissionStatusForm button[type="submit"]',
-      ) as HTMLButtonElement;
-
-      if (submitButton) {
-        // Submit the form via the enhance directive by clicking the button.
-        submitButton.click();
-      }
-    }, 10);
-  }
-
-  // FIXME: When customizing submission status, implement DB fetching for status options.
-  const submissionStatusOptions = submission_statuses.map((status) => {
-    const option = {
-      innerId: status.id,
-      innerName: status.status_name,
-      labelName: status.label_name,
-    };
-    return option;
-  });
+  let updatingDropdown: UpdatingDropdown;
 </script>
 
 <div class="flex items-center w-full space-x-1 text-left text-sm sm:text-md">
@@ -116,99 +62,12 @@
     <button
       type="button"
       class="flex-shrink-0 w-6 ml-auto"
-      onclick={dropdown.toggle}
+      onclick={() => updatingDropdown.toggle()}
       aria-label="Update submission for {selectedTaskResult.title}"
     >
       <EllipsisVertical class="w-4 h-4 mx-auto" />
     </button>
 
-    {@render dropdownList(selectedTaskResult)}
+    <UpdatingDropdown bind:this={updatingDropdown} {taskResult} {isLoggedIn} {onupdate} />
   </div>
-{/snippet}
-
-{#snippet dropdownList(selectedTaskResult: TaskResult)}
-  <div class="relative">
-    <Dropdown
-      {activeUrl}
-      {dropdownStatus}
-      {closeDropdown}
-      class="absolute w-24 z-20 left-auto right-0 mt-8"
-    >
-      <DropdownUl>
-        {#each submissionStatusOptions as submissionStatus}
-          <DropdownLi href="javascript:void(0)" onclick={() => handleSubmit(submissionStatus)}>
-            {submissionStatus.labelName}
-          </DropdownLi>
-        {/each}
-      </DropdownUl>
-    </Dropdown>
-
-    {#if showForm && selectedSubmissionStatus}
-      {@render submissionStatusForm(selectedTaskResult, selectedSubmissionStatus)}
-    {/if}
-  </div>
-{/snippet}
-
-{#snippet submissionStatusForm(
-  selectedTaskResult: TaskResult,
-  submissionStatus: { innerId: string; innerName: string; labelName: string },
-)}
-  <form
-    id="submissionStatusForm"
-    method="POST"
-    action="?/update"
-    style="display:none;"
-    use:enhance={({ formData, action, cancel }) => {
-      // Cancel the default form submission.
-      cancel();
-
-      // Submit data manually using fetch API.
-      fetch(action, {
-        method: 'POST',
-        body: formData,
-        headers: {
-          Accept: 'application/json',
-        },
-      })
-        .then((response) => response.json())
-        .then(() => {
-          const updatedTaskResult = {
-            ...taskResult,
-            status_name: submissionStatus.innerName,
-            status_id: submissionStatus.innerId,
-            submission_status_label_name: submissionStatus.labelName,
-            is_ac: submissionStatus.innerName === 'ac',
-            updated_at: new Date(),
-          };
-
-          onupdate(updatedTaskResult);
-        })
-        .catch((error) => {
-          console.error('Failed to submit task status:', error);
-        })
-        .finally(() => {
-          closeDropdown();
-          showForm = false;
-        });
-
-      // Do not change anything in SvelteKit.
-      return () => {};
-    }}
-  >
-    <!-- Task id -->
-    <InputFieldWrapper
-      inputFieldType="hidden"
-      inputFieldName="taskId"
-      inputValue={selectedTaskResult.task_id}
-    />
-
-    <!-- Submission status -->
-    <InputFieldWrapper
-      inputFieldType="hidden"
-      inputFieldName="submissionStatus"
-      inputValue={submissionStatus.innerName}
-    />
-
-    <button type="submit">Submit</button>
-  </form>
 {/snippet}
